requirements.txt
Flask==3.1.*
pytesseract==0.3.*
Pillow==11.*
pywebview==5.*
pyinstaller==6.*
